feat: split up telemetry into separate module, cleaned up app setup steps, pass in a testing flag for app setup
