+++ .env.template
@@ -1,3 +1,3 @@
 # DATABASE_URL=add_db_connection_here
-# LOGFIRE_ENVIRONMENT=environment name for otel
-# SERVICE_NAME=service name for otel
+# LOGFIRE_ENVIRONMENT=environment name for telemetry
+# SERVICE_NAME=service name for telemetry
